Fixed styling and displayed user in navbar
