+++ src/App.js
@@ -10,11 +10,10 @@ import Admin from "./components/admin/Admin";
 import Login from "./components/login/Login";
 import Navbar from "./components/nav/Navbar";
 import Homepage from "./components/customer/Homepage";
-import BookingWizard from './components/BookingWizard'
-
+import BookingWizard from "./components/BookingWizard";
 
 function App() {
-  const [user, loading] = useAuthState(auth)
+  const [user, loading] = useAuthState(auth);
   // state to toggle sidebar in admin
 
   const [toggleSidebar, setToggleSidebar] = useState(false);
@@ -23,26 +22,21 @@ function App() {
   return (
     <div className="container-fluid p-0">
       <Navbar setToggleSidebar={setToggleSidebar} />
-      <div className="row">
-      <div className="col-12">
-
       <Routes>
         <Route
           path="/admin"
           element={
             <ProtectedRoute isAllowed={!!user} loading={loading}>
-              <Admin toggleSidebar={toggleSidebar}/>
+              <Admin toggleSidebar={toggleSidebar} />
             </ProtectedRoute>
           }
-          />
+        />
         <Route path="/" element={<Homepage />} />
         <Route path="/login" element={<Login />} />
         <Route path="/booking/*" element={<BookingWizard />} />
       </Routes>
-          </div>
-          </div>
     </div>
   );
 }
 
-export default App
+export default App;
