+++ src/components/admin/Admin.jsx
@@ -1,43 +1,27 @@
 import React from "react";
 import { auth } from "../../firebase/client";
-import { logout } from "../../redux/authSlice";
-import { signOut } from "firebase/auth";
+
 import { useAuthState } from "react-firebase-hooks/auth";
-import { useDispatch } from "react-redux";
+
 import Sidebar from "./nav/Sidebar";
 
 const Admin = (props) => {
   const { toggleSidebar } = props;
   const [user, loading] = useAuthState(auth);
-  const dispatch = useDispatch();
 
-  const logoutHandler = () => {
-    signOut(auth);
-    dispatch(logout());
-  };
+  if (loading)
+    return <p className="text-center pt-2">Loading... please wait</p>;
+
   return (
-    <div className="row">
-      {loading ? (
-        <div className="col">
-          <p>Loading...please wait</p>
-        </div>
-      ) : (
-        <>
-          <div className="col-lg-2">
-            <Sidebar toggleSidebar={toggleSidebar} />
-          </div>
-          <div className="col-lg-10">
-            <div className="container text-center">
-              <p>Hello Admin: {user?.displayName}</p>
+    <main className="d-flex">
+      <div>
+        <Sidebar toggleSidebar={toggleSidebar} />
+      </div>
 
-              <button onClick={logoutHandler} className="btn btn-secondary">
-                Sign Out
-              </button>
-            </div>
-          </div>
-        </>
-      )}
-    </div>
+      <div className="container-fluid">
+        <p>Hello </p>
+      </div>
+    </main>
   );
 };
