+++ src/components/admin/nav/Sidebar.module.scss
@@ -27,5 +27,6 @@
     left: 0;
     display: block;
     z-index: 0;
+    width: 250px;
   }
 }
